integrations/alexa/smart-home-skill/smarthome/requirements.txt
paho-mqtt==1.5.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
paho-mqtt==1.5.0
requests==2.23.0
urllib3==1.25.8
